feat: attach ProductUpdateListener and process changes if the Product object is being updated
The ProductUpdateListener detects changes to the Product object and has an added check to make sure the Product is published.

Normally, I would have asked what parameters would make us want to create, update and either disable or something when a Product goes from a state of Published to Unpublished, but this is small-scale, so I'm hopeful this is acceptable for that purpose.

If anyone is reading this, kudos to you.